+++ lib/src/analyzer/rules/registry.dart
@@ -1,13 +1,21 @@
 import 'package:analysis_server_plugin/registry.dart';
 
-import 'hook_disallow_context_outside_build_rule.dart';
-import 'hook_in_control_flow_rule.dart';
-import 'hook_in_nested_function_rule.dart';
-import 'hook_requires_context_in_build_rule.dart';
+import 'disallow_context_outside_build_rule.dart';
+import 'effect_cleanup_requires_effect_rule.dart';
+import 'global_effect_requires_dispose_rule.dart';
+import 'no_hooks_in_control_flow_rule.dart';
+import 'no_hooks_in_nested_functions_rule.dart';
+import 'no_writes_in_computed_rule.dart';
+import 'require_context_in_build_rule.dart';
+import 'scope_dispose_requires_scope_rule.dart';
 
 void registerRules(PluginRegistry registry) {
-  registry.registerWarningRule(HookRequiresContextInBuildRule());
-  registry.registerWarningRule(HookDisallowContextOutsideBuildRule());
-  registry.registerWarningRule(HookInControlFlowRule());
-  registry.registerWarningRule(HookInNestedFunctionRule());
+  registry.registerWarningRule(RequireContextInBuildRule());
+  registry.registerWarningRule(DisallowContextOutsideBuildRule());
+  registry.registerWarningRule(NoHooksInControlFlowRule());
+  registry.registerWarningRule(NoHooksInNestedFunctionsRule());
+  registry.registerWarningRule(NoWritesInComputedRule());
+  registry.registerWarningRule(EffectCleanupRequiresEffectRule());
+  registry.registerWarningRule(ScopeDisposeRequiresScopeRule());
+  registry.registerWarningRule(GlobalEffectRequiresDisposeRule());
 }